Adding function to evaluate coverage probability
In doing so, discovered that there is an asymmetry in the number of
predicted wins when including home court.  I think this must be due to
the fact that some games are on neutral courts and this isn't
reflected in the way I am parsing the data.  For now, switching off
inclusion of home court, although did notice that using it
significantly improves confidence in the predictions while still
retaining good coverage.