Fix and improvements for home court tracking
Was incorrectly using a 0 value to represent away games in
evaluate_predicted_wins and evaluate_coverage_probability, and this
was the cause of the count mismatch.  After resolving this, the counts
line up.  Coverage confidence when accounting for home court is not as
much higher than without as it previously seemed.

Updates to parse home/away into 'LOC' column immediatey, so that the
logic is only used once.  Also added checking to verify that for each
game only one team is home, and it checks out.

Finally, switched default to include home court advantage.