Vectorize eval_move calculation
This dramatically speeds up the MCMC.  Now it is passable, although
not espeically fast.  Near the end of the season (~1200 total games),
and it runs at about 100 samples per second.

So far only limited testing with small sample sizes, but the resulting
rankings do look reasonable in comparison to the least squares results.