Adding capability to train/test on subsets of available data
Added option max_date_train to the League class and its children.
This sets up a TRAIN flag in each game data frame for whether or not
the game is used in training.  Was easy to update both least squares
and Bayesian methods to respect this training flag.  Then added an
optional flag to evaluate_predicted_wins for whether to exclude the
training cases.

For now updated the test codes to train through Jan 1, 2018 and
exclude these games in testing.  With current settings, appears that
the least squares model more accurately predicts games won.