Initial rework towards general rating framework
Reworking as general framework, not specific to nba_py.  For now, use
of nba_py is removed, but it could be added back in the future as a
data source.

Created new Team class for holding a data frame with team data.  The
intent is to provide class methods that can create team instances from
various data sources.  For now, starting with "from_hyper_table",
which uses a simple DataFrame format with TEAM_ID, GAME_ID, and PTS
columns, following what is available on Massey's site.  The League
class works similarly, with classmethods that can create a
league (list of teams) using appropriately formatted input data.

Modifying RatingSystem class to accept leage as argument instead of
inheriting from it.  Updating LeastSquares so that score_cap and
homecourt are options to the constructor.