Rework hyper games reader
Rework so that full team data is defined in the reader for each
individual team, as opposed to looping over the data as a
post-processing step at the league stage.  This is faster through the
use of vector operations, but it also means that team data is not in
an incomplete state prior to creation of a league.