Make flagging of scheduled games more robust
When reading and merging data, setting nan in only PTS was not enough
to flag a scheduled game.  Need to either check for a nan in either,
or place a nan in both PTS and OPP_PTS.  For safety, doing both.  This
was identified when testing with Massey data that included scheduled
games.