Refactor how predictions are done
Refactoring prediction functions to operate on games DataFrame input
instead of Team instances.  Moving code that gets the predictions into
ratingbase.  Also storing the full double-games table in ratingbase,
which can be used for getting predictions.

The ratings DataFrame has also been tweaked to use team_id as the
index instead of name with team_id as fallback.  This is necessary so
that the prediction code can access values in the ratings table in a
consistent way.