Add mode query parameter for rank vs rating
Idea is to support a display mode that shows ranks instead of ratings,
for e.g. strength of schedule.  The query parameter works, but no
links to it have been added yet.  One quirk is how to handle missing
values, which will currently get ranked at the top (for descending)).