Initial refactoring to eliminate Team class
Eliminate Team class and store team and game data in data frames
directly under League and RatingBase classes.  The tests pass, but
supllementary functionality such as writing to the database will still
need to be updated.

For now the RatingBase.teams attribute has been renamed df_teams to
clarify the change.  It might be changed back once the transition is
complete.