Initial take at implementation of weighted least squares
Weights are applied to the basis and observation values, and then the
least squares fitting is done as usual.  The challenges are in
calcualtion of R-squared and leave-one-out residuals.  An ad-hoc
attempt at computing an unbiased estimate of residual standard
deviation has been implemented, but it should be reviewed.  Believe
that the R-squared and leave-one-out calculations are correct, but
further review and testing would be good.

Adding new test case with weights.