Retain game indices
In constructor, when doubling games, retain the indices instead of
re-indexing.  This will facilitate testing, where it may be useful to
refer to games via the original indexing.

After this change, it was also necessary to add a .values attribute
when using a series to index into a numpy array (not 100% sure what
the issue is and why it worked before).