[web] Change off/def/SoS rating number format
Use 1 decimal for these, while leaving 2 decimals for main rating.
Considered switching to 1 decimal for main rating, but there is room
in the column for it (although with 1 decimal place, sorting does
still appear to "do the right thing" with matches, probably because it
falls back to using the first column (rank) as the secondary sort).