Add win_value option for MLE Wins metric
The idea is to allow a win to represent a fraction of the game.  This
is one approach for avoiding numerical issues with winless or lossless
teams.

Also changing the solution method of scipy.optimize.fixed_point to use
plain iteration, as I was seeing some failures (for college
basketball) when testing this out.

More work is still needed to optimize this for larger leagues such as
college basketball, which currently is somewhat slow even when using a
low tolerance (1e-1).