Treat all ratings as variables in MLE
Instead of constraining the last rating to 1, let all ratings float.
Even though there is an arbitrary scale in the solution, this appears
to result in much faster convergence (also noted by Massey).

Also switching the starting point to be based on win/loss ratio.  With
floating ratings, this appears to improve convergence vs just starting
with ones.  With these changes, finding the MLE for college basketball
is much more tractable.