Initial work to support unsaturated systems
Fall back to LAPACK DGELSS solver, which produces a minimum norm
solution.  Also adding a simple test.  Currently the normalization is
a little bit off, because it is finding the minimum norm solution with
the last equation removed, whereas we probably want to leave in the
last equation in this case (should be workable with some minor
tweaks).